+++ src/libromdata/Other/Wim.hpp
@@ -1,0 +1,19 @@
+/***************************************************************************
+ * ROM Properties Page shell extension. (libromdata)                       *
+ * Wim.hpp: Microsoft WIM header reader                                    *
+ *                                                                         *
+ * Copyright (c) 2023 by ecumber.                                          *
+ * Copyright (c) 2019-2023 by David Korth.                                 *
+ * SPDX-License-Identifier: GPL-2.0-or-later                               *
+ ***************************************************************************/
+
+#pragma once
+
+#include "librpbase/RomData.hpp"
+
+namespace LibRomData {
+
+ROMDATA_DECL_BEGIN(Wim)
+ROMDATA_DECL_END()
+
+}