+++ NEWS.md
@@ -1,5 +1,12 @@
 # Changes
 
+## v2.3 (released 2023/??/??)
+
+* New parsers:
+  * Wim: Microsoft Windows Images, used by the Windows installer starting with
+    Windows Vista. Contributed by @ecumber.
+    * Pull requests: #391, #392
+
 ## v2.2.1 (released 2023/07/30)
 
 * New parser features: